stdlib: update e2dk4r/stdlib
eea60e03658283c6f0a9d66c3ace53d3deb42e3f

+++ include/string_builder.h
@@ -55,7 +55,7 @@ static inline void
 StringBuilderAppendZeroTerminated(string_builder *stringBuilder, const char *src, u64 max)
 {
   struct string *outBuffer = stringBuilder->outBuffer;
-  struct string string = StringFromZeroTerminated((u8 *)src, max);
+  struct string string = StringFromZeroTerminated((u8 *)(u64)src, max);
   memcpy(outBuffer->value + stringBuilder->length, string.value, string.length);
   stringBuilder->length += string.length;
   debug_assert(stringBuilder->length <= outBuffer->length);
